Fixing cyclic import issue
